Added DHCP UDP IPV4 basic support

+++ net/net.c
@@ -38,7 +38,7 @@ void print_ethernet_packet(void *packet) {
   kprintf("\n");
 }
 
-u32 switch_endian16(u16 nb) { return (nb >> 8) | (nb << 8); }
+u16 switch_endian16(u16 nb) { return (nb >> 8) | (nb << 8); }
 
 u32 switch_endian32(u32 nb) {
   return ((nb >> 24) & 0xff) | ((nb << 8) & 0xff0000) | ((nb >> 8) & 0xff00) |
